[refactoring] step2: MineSweeper 클래스 다이어트
