+++ src/main/kotlin/minesweeper/main.kt
@@ -1,5 +1,7 @@
 package minesweeper
 
+import minesweeper.domain.MineSweeper
+import minesweeper.domain.MineSweeperWidth
 import minesweeper.domain.board.size.MineSweeperBoardSize
 import minesweeper.domain.mine.Mine
 import minesweeper.ui.InputView
@@ -15,9 +17,15 @@ fun main() {
     val mine = Mine(mineCount)
     val mineSweeperBoardSize = MineSweeperBoardSize(width = width, height = height)
 
-    val mineSweeper = MineSweeper(board = mineSweeperBoardSize, mine = mine)
+    val mineSweeper = MineSweeper(board = mineSweeperBoardSize)
     OutputView.startMineSweeper()
 
-    val mineSweeperBoard = mineSweeper.createBoard()
+    val notContainedMineCount = mineSweeperBoardSize.getBoardFullSize() - mine.mineCount
+    val notContainedMineList = MineSweeperWidth.newInstance(widthSize = notContainedMineCount)
+    val mineList = MineSweeperWidth.newInstance(widthSize = mine.mineCount, mineSweeperShape = mine.mineShape)
+
+    val mineSweeperList = (notContainedMineList + mineList).shuffled()
+
+    val mineSweeperBoard = mineSweeper.createBoard(mineSweeperList)
     OutputView.mineSweeperInitializePrinter(mineSweeperBoard)
 }
